Added a few loading screens
