+++ lib/pages/logged_in/profile_page/profile_widgets/profile_hero.dart
@@ -28,6 +28,7 @@ class ProfileHero extends StatelessWidget {
                 child: Center(
                   child: O3D(
                     src: avatarUrl,
+                    autoRotate: true,
                   ),
                 ),
               ),
