+++ lib/pages/logged_in/run_page/components/run_detail_and_stop.dart
@@ -1,5 +1,4 @@
 import 'dart:io';
-import 'package:firebase_core/firebase_core.dart';
 import 'package:runningapp/models/user.dart' as user;
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
